pyproject.toml
[project]
name = "agentic-compliance-analyst"
version = "0.1.0"
description = "Compliance analysis agent using LLMs"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "arize-phoenix>=12.31.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "groq>=1.0.0",
    "instructor>=1.14.4",
    "pydantic>=2.12.5",
    "pymupdf4llm[layout,ocr]>=0.2.9",
    "sse-starlette>=3.2.0",
    "streamlit>=1.53.1",
    "streamlit-lottie>=0.0.5",
    "streamlit-option-menu>=0.4.0",
    "tavily-python>=0.7.19",
    "trafilatura>=2.0.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
